requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin>=6.0.0
aiohttp>=3.8.0
websockets==11.0
python-dotenv>=1.0.0
tzlocal>=4.0.0
typing-extensions>=4.0.0
rich>=13.0.0
selenium>=4.0.0
webdriver-manager>=4.0.0
psutil>=5.9.0
loguru>=0.7.2
pydantic>=2.0.0
pandas>=2.0.0